requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
blis
certifi
cffi==1.17.1
charset-normalizer
click
cloudpathlib
colorama
confection
cryptography==43.0.3
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem
Deprecated==1.2.14
en_core_web_trf==3.8.0
exceptiongroup==1.2.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
Jinja2
joblib
langcodes
markdown-it-py
MarkupSafe
mdurl
microsoft-kiota-abstractions==1.6.2
microsoft-kiota-authentication-azure==1.6.2
microsoft-kiota-http==1.6.2
microsoft-kiota-serialization-form==1.6.2
microsoft-kiota-serialization-json==1.6.2
microsoft-kiota-serialization-multipart==1.6.2
microsoft-kiota-serialization-text==1.6.2
mkl-service==2.4.0
mkl_fft
mkl_random
mpmath==1.3.0
msal==1.31.0
msal-extensions==1.2.0
msgraph-core==1.1.6
msgraph-sdk==1.12.0
multidict==6.1.0
murmurhash
networkx==3.2.1
nltk
numpy==2.0.1
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
packaging
pendulum==3.0.0
portalocker==2.10.1
preshed
propcache==0.2.0
pycparser==2.22
pydantic
pydantic_core
Pygments
PyJWT==2.9.0
PySocks
python-dateutil==2.9.0.post0
pywin32==308
regex==2024.11.6
requests
rich
shellingham
six==1.16.0
smart-open
sniffio==1.3.1
spacy
spacy-curated-transformers==0.3.0
spacy-legacy
spacy-loggers
srsly
std-uritemplate==2.0.1
sympy==1.13.1
thinc
time-machine==2.16.0
torch==2.5.1
tqdm
typer
typing_extensions
tzdata==2024.2
urllib3
wasabi
weasel
Whoosh
win-inet-pton
wrapt==1.16.0
yarl==1.17.1
zipp==3.21.0
